Use uint64_t for DECODE_UNIT presentationTimeMs
This still holds a millisecond value sourced from RTP timestamp, but since we're changing the API anyway, now is a good time to future-proof this field.

+++ src/Limelight.h
@@ -167,7 +167,7 @@ typedef struct _DECODE_UNIT {
     // Presentation time in milliseconds with the epoch at the first captured frame.
     // This can be used to aid frame pacing or to drop old frames that were queued too
     // long prior to display.
-    unsigned int presentationTimeMs;
+    uint64_t presentationTimeMs;
 
     // Length of the entire buffer chain in bytes
     int fullLength;
